Fix: sessiond: Only open packets on unpublished streams
2d713ac8e changed the code of the "open packet" command to iterate on
every stream of a channel without regard as to their "published" state.

This causes a regression in the kernel trace-format test as unexpected
packets are being generated. More concretely, it causes extraneous
packets to be generated in the trace.

Looking at the commit message which introduced the command (04ed9e10),
this command was only ever meant to be used on channels that don't have
an output (essentially, snapshot sessions).

A filter is used to iterate only on unpublished streams, restoring the
initial behaviour.

Change-Id: I2ba80d1e9ec0d9d1b98b1c9da0908d32b52765f2
Signed-off-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>