tests: Rework notification generator in python
Observed issue
==============

A number of transient test aborts were observed in the following
tests:

* `tests/regression/tools/notification/test_notification_multi_app`
* `tests/regression/tools/notification/test_notification_ust_buffer_usage`
* `tests/regression/tools/notification/test_notification_capture`

The tests would hang indefinitely, typically with the consumer in a
consumption paused state.

In failing cases, I noticed the there were odd bash errors, such as:

    ok 90 - Subscribe to high condition
      ---
        duration_ms: 0.71351
      ...
    # Online CPUs: 35 61 75 81 91 101 115 129
    # Online CPUs: 35 61 75 81 91 101 115 129
    # Online CPUs: 35 61 75 81 91 101 115 129
    # Online CPUs: 35 61 75 81 91 101 115 129
    # Online CPUs: 35 61 75 81 91 101 115 129
    # Online CPUs: 35 61 75 81 91 101 115 129
    ok 91 - High notification received after intermediary communication
      ---
        duration_ms: 2015.405443
      ...
    tools/notification//../../..//utils/utils.sh: trap: line 2: unexpected EOF while looking for matching `)'
    # Online CPUs: 35 61 75 81 91 101 115 129
    # Test hangs from here on...

The errors seem to occur more frequently on armhf, arm64, and
riscv64. In these cases we are using the following bash versions:

* 5.2.15(1)-release (Debian bookworm)
* 5.2.37(1)-release (Debian trixie)

This bash error does occur on amd64, but much less frequently. From
iterative testing in one of my developement environments, it failed
once in ~1500 runs, instead of once in about ~50 runs on arm64.

When running the tests with `bash -x`, I noticed the error is omitted
after the `util_event_generator.sh` scripts were signaled, e.g.

    + err_output=/tmp/tmp.tgg57vIvYG
    + return 0
    + [[ -n y ]]
    + kill -s SIGUSR1 1598066
    + '[' '!' -f /tmp/tmp.test_notification_multi_app_state_file.izhR4x ']'
    + sleep 0.5
    ./tests/regression/tools/notification//../../..//utils/utils.sh: trap: line 2: unexpected EOF while looking for matching `)'
    ++ get_any_available_cpu
    +++ get_online_cpus
    +++ cpus=()

Cause
=====

This appears to be a racy bug in bash's signal handling. As of this
writing, the error doesn't seem to be occurring with bash 5.3.

From the bash CHANGES[1] file, there are a number of bug fixes related
to signal handling in 5.3; however, I haven't identified any bug
reports, list e-mails, or specific commits that appear to be directly
related.

Solution
========

While it may be possible to wait until all test runners have bash 5.3,
that wouldn't enter Debian until the next release (around 2027 at the
current cadence).

I have opted to convert the `util_event_generator.sh` script into
Python, in the hopes that it's signal handling isn't as fraught.

Known drawbacks
===============

None.

References
==========

[1]: https://github.com/bminor/bash/blob/a8a1c2fac029404d3f42cd39f5a20f24b6e4fe4b/CHANGES#L1

Change-Id: I06707a03c807c9515f70f5d8e5dacb053e11a131
Signed-off-by: Kienan Stewart <kstewart@efficios.com>
Signed-off-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>