Fix: common/optional: unaligned access when setting value
clang warns that setting the watchdog_timer value may result
in an unaligned access withing the LTTNG_OPTIONAL_SET macro.

To work around that, the macro invokes two functions which
defer to memset/memcpy (details::lttng_optional_set), one of
which is overloaded for r-values which require a local copy
to be performed before invoking memcpy.

Change-Id: Icb6996fa1e13a16463650abc92bb16285716b740
Signed-off-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>