Tests: stall-buffer: initialize first packet before running
Following https://github.com/lttng/lttng-ust/commit/81b931f1,
the buffer stall tests fail on the CI.

That's essentially because the LTTng-UST change introduces an
unexpected new "state" to test for. Indeed, with the LTTng-UST
change applied, the buffer's first packet isn't initialized.

The tests will be adapted to test this new axis, but for the
moment emiting a single event before running the scenarios
puts them back in the state that they expect to see.

Change-Id: I97851ef31dafa7878364844b0cd27d7c0c1eca20
Signed-off-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>