Tests: stall-buffer: cover lazy packet allocation states
The simple tests were assuming that the very first packet header was
allocated prior any events. This results in wrong assumptions about
packets and events numbers present in the final traces in some
scenarios.

Fix this by considering that the very first packet header is now lazily
allocated on the very first event.

Change-Id: Ifc42f46f5ef89be8bfcaf034160218397b76f4e3
Signed-off-by: Olivier Dion <odion@efficios.com>
Signed-off-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>