tests: stall-buffer: Handle numeric suffix of testpoints
Testpoints in LTTng-UST are now generated with a numeric suffix that is
unique within a compilation unit. This is to avoid overwriting
testpoints duplicated due to the inlining of functions.

To install a breakpoint on a testpoint T, all testpoints matching TN,
where N is a numeric value, must have a breakpoint installed on them.

By using GDB's Python API, we can list the symbols found in objects
loaded in memory that match the name `lttng-ust.so'. Only symbols
matching `lttng_ust_testpoint' are considered.

Since we want to keep the temporary breakpoint semantic, all installed
breakpoints are removed after scheduling a producer using GDB's `delete`
command.

Finally, since listing the testpoint symbols can be slow, the timeout of
GDB is 30 seconds instead 5.

Depends-on: lttng-ust: I50f532afff02d62413dc65d2d92d7ea60776f4a2
Change-Id: Ifc6d1b18999844a96169268225b5ccfbd2aba058
Signed-off-by: Olivier Dion <odion@efficios.com>
Signed-off-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>