tests: Strong reproducibility of stall buffers
There are two components that are affecting the reproducibility of the
stall buffer tests.

First, the watchdog timer period. Depending on the system workload, the
watchdog timer would be triggered at certain times, producing a spurious
empty packet in the final trace of some stall scenarios. Expectations of
problematic scenarios could have been changed, but it is better to
simply disable the watchdog timer, letting the stall fixup on session
destroy only.

Second, on system with heavy workload, spurious timeout failure can
happen. This typically happens with the GDB process (the producers
scheduler) because it has to dynamically list the testpoints in the UST
shared library. Therefore, to minimize the chance of this happening, run
the stall buffer tests serially instead of in parallel.

Change-Id: I017aec4cb8a938e4891196a741fa5f4b1e5d88d0
Signed-off-by: Olivier Dion <odion@efficios.com>
Signed-off-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>