tests/stall-buffer: Set exit breakpoint on scheduling
Previously, the `tbreak' command was used to insert breakpoints to
testpoints.  However, commit 4b72aa7e7c08cb45cedf3722e0801847e26beecf
introduces matching of testpoints with numeric suffix to handle
duplicated testpoints in inlined functions.

The temporary semantic of `tbreak' was emulated by deleting all
breakpoints after a producer has stopped.  But this also removed the
breakpoint to `exit()', resulting in impossible scenarios to fail
because GDB could not kill all producers.

Fix this by installing a breakpoint on `exit()' before scheduling a
producer, if the scenario is marked as "might be impossible".

Change-Id: I2bcc5ff301ebe5cf442b77ecd304ef63963925c1
Signed-off-by: Olivier Dion <odion@efficios.com>
Signed-off-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>