tests: Add LTTNG_CURRENT_TEST env var
When a test run leaves stray processes it can be hard to pinpoint from
which specific test script they originate. To make this easier, add the
LTTNG_CURRENT_TEST environment variable set to the current test script
in the test runners.

When encountering a stray process, its originating test script can be
determined by running:

  strings /proc/$PID/environ | grep LTTNG_CURRENT_TEST

Where $PID is the pid of the stray process.

Change-Id: I74503c74749fb1657c4e16ca6c7b9fc980c2e567
Signed-off-by: Michael Jeanson <mjeanson@efficios.com>
Signed-off-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>