Fix: `user_space_consumer_channel_keys` skips some per-uid channels
Observed issue
==============

When tracing a session configured with per-uid buffer ownership policy
and multiple owners have active applications, `lttng clear` does not
clear all the data in the buffers.

```
$ lttng create --snapshot
$ lttng enable-event -u --all
$ lttng start

$ gen_ust -i 10
$ sudo -u anotheruser gen_ust -i 10

$ lttng snapshot record /tmp/a
$ babeltrace2 /tmp/a | wc -l
20

$ lttng clear
$ lttng snapshot record /tmp/b
$ babeltrace2 /tmp/b | wc -l
10
```

Cause
=====

When advancing the `user_space_consumer_channel_keys` iterator for
per-uid applications, there are cases where a double-advance occurs.

Solution
========

Use an else statement to avoid double-advancing the iterator.

Known drawbacks
===============

It looks to me like the case with a channel filter that is a metadata
channel would cause an assert.

Change-Id: Ieadb8e52a1aaf6230e3845e77a95c9f5af50c3ed
Signed-off-by: Kienan Stewart <kstewart@efficios.com>
Signed-off-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>