Fix: sessiond: validate iterator in get-channel-memory-usage
Coverity reports INVALIDATE_ITERATOR (CID 1635269) indicating that an
iterator from std::unordered_map::find() was being dereferenced without
checking if it was valid (i.e., not equal to end()).

In practice this can't happen since the descriptions are populated just
before, but there's no material cost to checking and logging the error
if a bugs finds its way there at some point.

This fix adds a check to ensure the iterator is valid before dereferencing
it, and throws an error if the consumer channel key is not found in the
channel descriptions map.

Change-Id: Id7c1c7db60b391cf6c19da298cc6956aaa53a43d
Signed-off-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>