Fix: lttng-reclaim-memory: check return value of mi_lttng_writer_command_close
Coverity reported CHECKED_RETURN (CID 1635265) indicating that
mi_lttng_writer_command_close() was being called without checking its
return value, whereas it is checked in 25 out of 26 other call sites.

This fix adds a check for the return value within the scope_exit guard
and logs an error if the function fails, since the guard is noexcept
and cannot throw.

Change-Id: I743638fafd9f3799f83220c32450922c931c9be8
Signed-off-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>