Add `src/bin/lttng/commands/list-wrappers.hpp`: C++ wrappers to list
This patch adds `src/bin/lttng/commands/list-wrappers.hpp` which
contains new C++ wrappers of the required parts of the liblttng-ctl C
API to implement the `lttng list` command.

The class hierarchy, within `lttng::cli`, is:

    tracepoint_field

    instrumentation_point
      kernel_tracepoint
      kernel_syscall
      ust_tracepoint
      java_python_logger

    instrumentation_point_set
      kernel_tracepoint_set
      kernel_syscall_set
      ust_tracepoint_set
      java_python_logger_set

    uprobe_location
      uprobe_function_location
      uprobe_tracepoint_location

    event_rule
      kernel_tracepoint_event_rule
      linux_syscall_event_rule
      linux_kprobe_event_rule
      linux_uprobe_event_rule
      ust_tracepoint_or_java_python_logger_event_rule
        ust_tracepoint_event_rule
        java_python_logger_event_rule

    event_rule_set

    data_stream_info
    data_stream_info_set
    data_stream_info_sets

    channel
      kernel_channel
      ust_or_java_python_channel
        ust_channel
        java_python_channel

    channel_set

    domain
      kernel_domain
      ust_domain
      java_python_domain

    domain_set

    rotation_schedule
      rotation_schedule_size
      rotation_schedule_periodic

    rotation_schedule_set

    snapshot_output
    process_attr_value
    process_attr_tracker
    session
    session_list

I'm using the `java_python` terminology to avoid confusion with "true"
LTTng-UST stuff and because the "agent" term is confusing.

The wrappers are as lightweight as possible, holding liblttng-ctl
pointers and calling library C functions as needed. Access the
liblttng-ctl object pointer of any wrapper with its lib() method. That
being said, I chose to put all the arrays of objects into ordered sets
(still of lightweight wrappers) to sort the items automatically and keep
a deterministic output, except for `session_list` of which the order, as
is, is significant (the first item is the newest recording session).

Although you may create any wrapper from one or more liblttng-ctl
pointers, the main entry points are the constructors of the
following classes:

• `kernel_tracepoint_set`
• `kernel_syscall_set`
• `ust_tracepoint_set`
• `java_python_logger_set`
• `session_list`

Those objects take care themselves of the lifetime of their
wrapped pointer(s).

All the classes holding a set or a list (with a name usually ending with
`_set` or `_list`) offer the convenient begin(), end(), size(), and
is_empty() methods.

Signed-off-by: Philippe Proulx <eeppeliteloop@gmail.com>
Change-Id: Idceca5719ba2cfb6596c66f55304bec34e92008f
Signed-off-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>