lttng-list(1): use a common configuration object for MI and human
This patch adds the `list_cmd_config` structure to `list-common.hpp` to
hold the list command configuration for both list_mi() and the new
list_human() (which is what used to be cmd_list(), but without the CLI
argument parsing).

cmd_list() now takes care of creating a `list_cmd_config` instance and
dispatching to either list_mi() or list_human() with it.

Both `list-mi.cpp` and the new `list-human.cpp` set a global
configuration variable which all their local functions read to operate.

Signed-off-by: Philippe Proulx <eeppeliteloop@gmail.com>
Change-Id: I9ad9225645d89e8b9de8ba3dc92e206bc746c110
Signed-off-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>