sessiond: Gate channels memory reclamation
Memory reclamation of channels is only valid if the shared-memory path
of the session owning the channels is on a file-system that support the
`MADV_REMOVE' advice for `madvise(2)'.  This check is done when creating
a UST session and is memoized for later sessions.

Furthermore, automatic memory reclamation of channels in the kernel
domain is not supported.

Finally, automatic memory reclamation of consumed memory does not apply
to:

  - Channels in a snapshot mode session
  - Channels in a session with no output

since by definition, the data is never consumed.

Therefore, remove the check for incompatible `--overwrite` option with
`--auto-reclaim-memory-consumed` from the client side of the
enable-channel command.  Instead, do all the check on the session-daemon
side since the session and system attributes are needed to determine if
automatic reclamation of consumed memory is possible for the wanted
channel.

Change-Id: I1f991f686d9d1da7500465e34a35529f220ca0b2
Signed-off-by: Olivier Dion <odion@efficios.com>
Signed-off-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>