Fix: relayd-grouping/test_ust: Remove nowait parameter for destroy after rotate
The documentation of lttng-destroy(1) is relatively unclear on whether
a sequence of `rotation` followed by `stop`, and then `destroy nowait`
yields to a stable trace.

I suspect that the rename of the last active chunk to the completed chunks
archive happens concurrently with the babeltrace execution, causing the
flakiness.

Remove the nowait parameter on destroy within tests that perform session
rotation.

Depends-on: lttng-ust: I577dc7bdfb467b5d172aff7a18f16775707833e0
Signed-off-by: Mathieu Desnoyers <mathieu.desnoyers@efficios.com>
Change-Id: Ib6f5fca83297b752642e44f5e4e4cf0cfb55a0e2