ust-consumer: Report channel memory usage with mincore(2)
Reporting the memory usage with `fstat(2)` can sometimes be erroneous
depending on the file-system implementation. A closer approximation
can be obtained using the `mincore(2)` syscall.

It is available since Linux 2.3.99pre1 and glibc 2.2, but is not
specified by POSIX.1. Therefore, for other operating systems, fall back
to the `fstat(2)` approximation.

Furthermore, if for some reason the `mincore(2)` syscall does not work,
fall back to the `fstat(2)` approximation. This check is only done once
and is memoized.

Depends-on: lttng-ust: If384db6b75af4c5bf2be031f07c6682b21ada2e4
Change-Id: I8277eeebcb73deea22284e176193f6a232debea7
Signed-off-by: Olivier Dion <odion@efficios.com>
Signed-off-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>