Fix: relayd: signal parent when daemon is actually ready
Introduced in 2.14, the '--sig-parent / -S' command line switch adds the
option to signal the parent of the relay daemon when it's ready to
accept connections.

The current behavior however sends this signal when all threads have
been created but before they are ready, change this to send the signal
when the threads have completed their initialization.

Change-Id: Id3816756e839c1287d4f73e0e320e6b4480d4713
Signed-off-by: Michael Jeanson <mjeanson@efficios.com>
Signed-off-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>