Revert "doc/man/Makefile: Mimic mandb(5) path hierarchy"
This patch introduces directory symlinks in the build tree that points
to their parent directory resulting in a loop when recursively walking
the directory tree.

You might argue that programs should be aware of this possibility and
break the loop when it arises, and I would agree, unfortunatly such
non-optimal programs exists and one of them is the artifact archiver of
Jenkins under certain circumstances.

A cleaner solution to the original problem this patch adresses would be
to generate our manpages in the proper subdirectories matching their
section number in mandb.

This reverts commit 46d51dbdc8b9c93871c34416006cf7f28b987c90.

Change-Id: I4f28d51a91bf0ad225e26e6ad9cfdb315a14f5d2
Signed-off-by: Michael Jeanson <mjeanson@efficios.com>
Signed-off-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>