Fix: relayd: Initialize stream_id when live inactivity beacon is received
Observed Issue
==============

Since LTTng-UST commit 81b931f1da86 ("Don't produce an initial packet header on ring-buffer creation"),
the following scenario causes Babeltrace 2 to assert:

lttng create --live
lttng enable-event -u all
lttng start
./gen-ust-events -i 10
TARGET=$(babeltrace2 -i lttng-live net://127.0.0.1:5344 | cut -d ' ' -f 1)
babeltrace2 -i lttng-live ${TARGET} > $TMP &
sleep 2
./gen-ust-events -i 15

Babeltrace 2 dies with this assert:

 (╯°□°)╯︵ ┻━┻  plugins/ctf/lttng-live/viewer-connection.cpp:1381: lttng_live_get_next_index(): Assertion `stream->ctf_stream_class_id.value == ctf_stream_class_id` failed.
Aborted

This also causes tools/live/test_lttng_ust_two_apps.py (run as root with
destructive tests for per-uid buffers) to fail with the same babeltrace
assertion.

Cause
=====

When a live beacon is received by the relay daemon before any stream
index, it fails to set the stream_id and leaves it to -1. Then, if the
live viewer is allowed to observe an inactivity status for that stream
before the stream receives an index which sets the stream_id, it
receives a stream_id=-1ULL for that stream. Subsequent activity for that
stream will then send indexes with a stream_id which does not match the
one initially observed, thus triggering the assertion.

Solution
========

Initialize the stream id when the live beacon is received for a stream
if it was not already initialized.

Signed-off-by: Mathieu Desnoyers <mathieu.desnoyers@efficios.com>
Change-Id: I0a0e12e782f7383c90d819d7580e4a2548726c8e
Signed-off-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>