tests: Fix spurious failure of stall-buffer timer
When the watchdog timer is disabled, the only thing that can force a
stalled-buffers fixup is when the session enters a quiescent state,
i.e. stopping the session.

Before this fix, only an asynchronous rotation was done before reading
the trace. This is okay when the watchdog timer is enabled because the
trace will eventually get fixed. However, when disabled, it is necessary
for the session to be stopped before reading the trace.

Change-Id: I9dff8d81955a00428ff11cd57da8c687a6b6a49b
Signed-off-by: Olivier Dion <odion@efficios.com>
Signed-off-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>