sessiond: session.hpp: add `ltt_session::trace_format`
Add a `trace_format` member to `ltt_session` to track the trace format
of a recording session.

The default value is `LTTNG_TRACE_FORMAT_CTF_1_8` for the moment,
therefore no functional changes intended until changed.

This patch only adds the member; usage is reserved to
subsequent patches.

This patch is part of an effort to add per-recording session trace
format configuration.

Signed-off-by: Philippe Proulx <eeppeliteloop@gmail.com>
Change-Id: Ic85f31d13266cca94f3509f320dfa0170e6941b3
Signed-off-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>