common: relayd: add relayd_create_session_2_15()
Add a function to send a version v2.15 "create session" command to the
relay daemon.

This function uses the relayd_create_session_2_11_base_populate() helper
to populate the base fields, then sets the trace format field specific
to version 2.15.

This is part of an effort to add per-recording session trace
format configuration.

Signed-off-by: Philippe Proulx <eeppeliteloop@gmail.com>
Change-Id: I71c64a9207713ac8e5e33ded04167ca54518697c
Signed-off-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>