sessiond: save.cpp: save trace format in session configuration
In save_session(), write the trace format of the recording session to
the XML configuration file.

The saved value is always concrete (`CTF_1_8` or `CTF_2`), not
`DEFAULT` (doesn't exist), since the session daemon will resolve the
default trace format at session creation time.

This is part of an effort to add per-recording session trace
format configuration.

Signed-off-by: Philippe Proulx <eeppeliteloop@gmail.com>
Change-Id: Ie9d5e97ef73356e1952da4e0a725008909acfa31
Signed-off-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>