sessiond: session-config.cpp: parse trace format when loading session config.
In process_session_node(), parse the `trace_format` XML element and pass
the corresponding trace format to create_session()
and create_snapshot_session().

When loading a pre-v2.15 configuration file without a `trace_format`
element, default to CTF 1.8 for backward compatibility.

This is part of an effort to add per-recording session trace
format configuration.

Signed-off-by: Philippe Proulx <eeppeliteloop@gmail.com>
Change-Id: I821a1c2392eea4f993c036567c37ba124957ab8b
Signed-off-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>