tests: test_save: add test_save_trace_format_one() helper
Add a helper function to test creating a recording session with a
specific trace format, saving it, and verifying the saved configuration
contains the correct trace format.

A subsequent patch will add a test which uses this.

This is part of an effort to add per-recording session trace
format configuration.

Signed-off-by: Philippe Proulx <eeppeliteloop@gmail.com>
Change-Id: I5a416d271538d47193307a1154f9b70c582761f8
Signed-off-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>