liblttng-ctl: add public `LTTNG_ERR_UNSUPPORTED_TRACE_FORMAT` error code
Add a new public error code to indicate that a trace format
isn't supported.

This will be used to report when a CTF 2 recording session tries to use
a pre-v2.15 relay daemon or pre-v2.15 kernel tracer modules.

Documentation left for a subsequent patch.

This patch is part of an effort to add per-recording session trace
format configuration.

Signed-off-by: Philippe Proulx <eeppeliteloop@gmail.com>
Change-Id: I2858bc53b3e5c63d828bbcdf5d9bc8f02e6731e8
Signed-off-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>