liblttng-ctl: add public lttng_session_descriptor_set_trace_format()
Add a public function to set the trace format of a session descriptor.

Users can call this before lttng_create_session_ext() to specify a
specific trace format for the recording session instead
of `LTTNG_TRACE_FORMAT_DEFAULT`.

Documentation left for a subsequent patch.

This is part of an effort to add per-recording session trace
format configuration.

Signed-off-by: Philippe Proulx <eeppeliteloop@gmail.com>
Change-Id: I42e21fa19df25da964ed3af10b08e643db5c734f
Signed-off-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>